backend/requirements.txt
# Backend API Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0

# Multi-Agent System
langgraph>=0.2.16
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
duckduckgo-search>=6.0.0
wikipedia>=1.4.0

# Persistence (SqliteSaver)
aiosqlite>=0.19.0

# Monitoring (Langfuse)
langfuse>=2.0.0

# Additional dependencies
openai>=1.0.0
